Moved object instantiation into init action
Moved the instantiation and global $facebook into the "init" action
which should solve the function not being defined before execution.

This is covered in http://core.trac.wordpress.org/ticket/14024

Also added 'admin_footer' hook so fb-root is created on admin pages as
well.  If this was intentional I apologize.